LayoutTests/security/contentSecurityPolicy/object-src-none-blocks-quicktime-plugin-replacement.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="object-src 'none'">
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

if (window.internals)
    window.internals.settings.setQuickTimePluginReplacementEnabled(true);

function done()
{
    if (window.testRunner)
        testRunner.notifyDone();
}

document.addEventListener("securitypolicyviolation", done, false);
</script>
</head>
<body>
<embed id="embed" width="640" height="480" src="../../plugins/resources/yellow.mov" qtsrc="../../plugins/resources/orange.mov">
</body>
</html>
